Progress #276 Reimplemented BeeHive world generation
Bee hives are now generated by using the center area of the given chunk
that has passed the rarity check set by worldgenBeeHiveRarity (default
10%). Also, by default we will attempt to spawn maxBeeHivesPerChunk
(default 4) in an area set by worldGenBeeHiveRadius around the center
of the chunk.

We have also added additional logging to beeHive generation. If you set
logLevel to debug in the growthcraft_bees.cfg it will log additional
details of which chunks are being checked and where the beeHives are
being generated.