Refactoring Implement sharable BlockColorHandler GrowthcraftBlockLog
In order to reduce code duplication, added BlockColorHandler and a
sharable Core class for GrowthcraftBlockLog.

BlockColorHandler - Took the colorizing code from BlockAppleLeaves and
made it a static method that can be used for all BlockLeaves in
Growthcraft.

GrowthcraftBlockLog - As we continue to add new trees, it only makes
sense to use a standard BlockLog class.