Fixed #292 Improvements from feedback in discord
Added a configuration for GrowthcraftMilk to allow for the customization
of the drop rate of stomach and thistle. Thistle drops at a rate of 2%
and stomach drops at a rate of 50%.

Added a configuration for limiting the dropping of stomach to only baby
cows. By default, stomach only drops from baby cows at a 50% chance.

Fixed #294 Implement inheritable super class for Configuration

Implemented the GrowthcraftConfiguration as the super class for
initializing the general config for mods.