Progress #309 Refactor BlockRopeKnot blocks to be more flexible
Changed the way that rope knot blocks are instantiated. When using a
rope on a fence it will check the ropeKnotFence ore dictionary for
an entry. If the fence is in the dictionary, then it will replace
the fence post with the rope knot equivalent.

In order to maintain the generic oak rope knot fences, left a generic
constructor that rolls up to the new method.

With this commit, you can no longer create the old generic rope knot
fences.