Fixed #323 Yeast now generates in culture jars
When we normalized the device processing across all of our devices, we
didn't realize that the Culture Jar had a special update that processed
the generation of yeast that was different from all of our other
devices.

To fix the generation issue, added an override to update for the
yeast generator class and then continued by calling the super update
so that the device progress gets updated.